+++ trunk/desmume/src/windows/replay.cpp
@@ -37,8 +37,8 @@ void Describe(HWND hwndDlg)
 	u32 num_frames = md.records.size();
 
 	double tempCount = num_frames / (33513982.0/6/355/263);
-	int num_seconds = tempCount;
-	int fraction = (tempCount - num_seconds) * 100;
+	int num_seconds = (int)tempCount;
+	int fraction = ((int)tempCount - num_seconds) * 100;
 	int seconds = num_seconds % 60;
 	int minutes = (num_seconds / 60) % 60;
 	int hours = (num_seconds / 60 / 60) % 60;
